% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nbytes}
\alias{nbytes}
\title{Get the number of bytes.}
\usage{
nbytes(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
An integer vector giving the number of bytes in each string.
}
\description{
Get the number of bytes in each string of a character vector.
}
\details{
Simple wrapper around \code{nchar(x, type = "bytes")}.

\code{NA} values are preserved in the output.
}
\examples{
nbytes(c("a", "ab", "abc"))
nbytes("caf\u00e9")
}
