% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape.r
\name{re_escape}
\alias{re_escape}
\title{Escape regex special characters}
\usage{
re_escape(string)
}
\arguments{
\item{string}{Character vector.}
}
\value{
A character vector the same length as \code{string}.
}
\description{
\code{re_escape()} returns a character vector with all regex
special characters escaped.
}
\details{
Inputs must be valid UTF-8 to work with Rust's regex engine.
Any non-UTF-8 elements will result in an error.

\code{NA} values are preserved and returned as \code{NA} in the
output - but note \code{NA} is not a valid pattern for rure.
}
\examples{
re_escape(".")
}
