% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.r
\name{re_where}
\alias{re_where}
\alias{re_set_where}
\alias{re_set_where_each}
\title{Find elements where regex match occurs}
\usage{
re_where(string, pattern)

re_set_where(string, patterns)

re_set_where_each(string, patterns)
}
\arguments{
\item{string}{Character vector.}

\item{pattern}{Pattern to look for (single string).}

\item{patterns}{Patterns to look for (character vector).}
}
\value{
For \code{re_set_where()} and \code{re_set_where_each()}, an
integer vector. For \code{re_set_where_each()}, a list the length
of \code{patterns} with each element containing an integer vector.
}
\description{
Return the indices of elements in \code{string} that match
regex patterns.
}
\details{
\code{re_where()} returns the indices of elements in \code{string}
that match \code{pattern}.

\code{re_set_where()} does the same for a set of patterns,
returning indices of elements in \code{string} where
\strong{any} of the patterns match.

\code{re_set_where_each()} returns a list of integer vectors,
where each list element corresponds to a pattern in \code{patterns}
and is filled with the indices of elements in \code{string}
that match the corresponding pattern.

No matches will result in an empty integer vector. For
\code{re_set_where_each()}, this will result in a list of empty
integer vectors.

Patterns must be valid UTF-8 to work with Rust's regex engine.
Any non-UTF-8 patterns will result in an error. \code{string} may contain
arbitrary bytes but ASCII compatible text is more useful, and UTF-8
is more useful still. Other text encodings are not supported.
}
\examples{
fruit <- c("apple", "banana", "pear", "pineapple")
re_where(fruit, "a")
re_where(fruit, "^a")
re_where(fruit, "a$")
re_where(fruit, "b")
re_where(fruit, "[aeiou]")

re_set_where(fruit, c("a", "e"))
re_set_where_each(fruit, c("a", "e"))
}
\seealso{
\link{re_detect} for logical return values and \link{re_find} for match
locations.
}
