% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.r
\name{re_detect}
\alias{re_detect}
\alias{re_set_detect}
\alias{re_set_detect_each}
\title{Detect regex matches}
\usage{
re_detect(string, pattern)

re_set_detect(string, patterns)

re_set_detect_each(string, patterns)
}
\arguments{
\item{string}{Character vector.}

\item{pattern}{Pattern to look for (single string).}

\item{patterns}{Patterns to look for (character vector).}
}
\value{
For \code{re_detect()} and \code{re_set_detect()}, a logical
vector the same length as \code{string}. For
\code{re_set_detect_each()}, a logical matrix with one row per
element of \code{string} and one column per pattern in \code{patterns}.
}
\description{
Return logical values indicating where regex patterns match
the input strings.
}
\details{
\code{re_detect()} returns a logical vector with \code{TRUE} for
each element of \code{string} that matches \code{pattern} and \code{FALSE}
otherwise.

\code{re_set_detect()} does the same for a set of
patterns, returning \code{TRUE} if \strong{any} of the patterns match.

\code{re_set_detect_each()} returns a logical matrix with one
row per element of \code{string} and one column per pattern
in \code{patterns} - showing which patterns match each string.

\code{NA} values in \code{string} will result in \code{NA} in the output.

Patterns must be valid UTF-8 to work with Rust's regex engine.
Any non-UTF-8 patterns will result in an error.
}
\examples{
fruit <- c("apple", "banana", "pear", "pineapple")
re_detect(fruit, "a")
re_detect(fruit, "^a")
re_detect(fruit, "a$")
re_detect(fruit, "b")
re_detect(fruit, "[aeiou]")

re_set_detect(fruit, c("a", "e"))
re_set_detect_each(fruit, c("a", "e"))
}
\seealso{
\link{re_find} for index return values.
}
