% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.r
\name{re_find}
\alias{re_find}
\alias{re_set_find}
\alias{re_set_find_each}
\title{Find regex matches}
\usage{
re_find(string, pattern)

re_set_find(string, patterns)

re_set_find_each(string, patterns)
}
\arguments{
\item{string}{Character vector.}

\item{pattern}{Pattern to look for (single string).}

\item{patterns}{Patterns to look for (character vector).}
}
\value{
For \code{re_set_find()} and \code{re_set_find_each()}, an
integer vector. For \code{re_set_find_each()}, a list the length
of \code{patterns} with each element containing an integer vector.
}
\description{
Return the indices of elements in \code{string} that match
regex patterns.
}
\details{
\code{re_find()} returns the indices of elements in \code{string}
that match \code{pattern}.

\code{re_set_find()} does the same for a set of patterns,
returning indices of elements in \code{string} where
\strong{any} of the patterns match.

\code{re_set_find_each()} returns a list of integer vectors,
where each list element corresponds to a pattern in \code{patterns}
and is filled with the indices of elements in \code{string}
that match the corresponding pattern.

No matches will result in an empty integer vector. For
\code{re_set_find_each()}, this will result in a list of empty
integer vectors.
}
\examples{
fruit <- c("apple", "banana", "pear", "pineapple")
re_find(fruit, "a")
re_find(fruit, "^a")
re_find(fruit, "a$")
re_find(fruit, "b")
re_find(fruit, "[aeiou]")

re_set_find(fruit, c("a", "e"))
re_set_find_each(fruit, c("a", "e"))
}
\seealso{
\link{re_detect} for logical return values.
}
