% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.r
\name{re_find}
\alias{re_find}
\title{Find locations of regex matches}
\usage{
re_find(string, pattern, start = 1L)
}
\arguments{
\item{string}{Character vector.}

\item{pattern}{Pattern to look for (single string).}

\item{start}{Byte offset at which to start searching (\code{1}-based).
Default is \code{1} (start at the beginning of each string).}
}
\value{
For \code{re_find()}, a (integer) matrix with two columns:
\code{start} and \code{end}. Each row corresponds to an element of \code{string}
and contains the \code{start} and \code{end} offsets of the match.
If no match is found, the row contains \code{NA} values.
}
\description{
Return the start and end offsets (in bytes)
of regex matches within each element of \code{string}.
}
\details{
\code{string} may contain arbitrary bytes but ASCII compatible
text is more useful, and UTF-8 is more useful still. Other
text encodings are not supported.

No match will result in \code{NA} values in \code{start}
and \code{end}. The same will occur for elements of \code{string} that
are shorter than the specified \code{start} offset.

Patterns must be valid UTF-8 to work with Rust's regex engine.
Any non-UTF-8 patterns will result in an error.
}
\examples{
fruit <- c("apple", "banana", "pear", "pineapple")
re_find(fruit, "ap")
re_find(fruit, "ap", start = 2)
}
\seealso{
\link{re_detect} for logical return values.
}
